jobs/monitoring: Add Prometheus
The change adds the Prometheus job. I had to go with certain assumptions
regarding the host constraint, as well as lack of volumes, but I hope
this serves as the base for running the monitoring stack on Nomad
itself.

The Prometheus instance is on purpose left unexposed, so that we can use
Grafana and additional authentication in that layer later.