ansible: Deploy nomad jobs using ansible
This is an untested change because I don't have the same environment as
Redbrick, but I think it would be beneficial to deploy the jobs using
Ansible. This means that it can be automated in the future.

One drawback of this approach is that we don't plan ahead of time, but
that should be done as part of a pull request anyways.