Add job for a nova-timetable user project
This file directory structure can be used for 4th year projects
(user-projects/(third/fourth)-year-project)

I can't validate this, but checkout this branch on one of the hosts and
run `nomad job validate <path>` to ensure the config is okay, followed
by `nomad job plan <path>` to plan it. Run the job at that point using
the check-id. Once you're happy that it runs, return to the master
branch and merge this PR.

nova@redbrick.dcu.ie requested this to be deployed.

A DNS entry will need to be made before planning to ensure that the cert
can be generated by traefik.